Allow to customize the table name (#361)
* Add an option to specify the table name to use

* Make the migration use the table name from the config

* Make Activity model use table name from the config

The `__construct()` method is needed because PHP won’t allow to call the `config()` function when defining the `$table` property on the model.

We keep the initialization statement of the `$table` property to ensure that it will have a protected visibility as intended. Without that statement, the constructor would create this property as a public one.